requirements.txt
accelerate==1.12.0
json-repair==0.54.2
jupyter==1.1.1
matplotlib==3.10.7
pandas==2.3.3
qwen-vl-utils==0.0.14
scikit-learn==1.8.0
torch-c-dlpack-ext==0.1.3
rich==14.2.0
plotext==5.3.2
hydra-core==1.3.2
wandb[media]==0.23.1
seaborn==0.13.2
pydantic==2.12.5
